+++ tests/testthat/test-carrier.R
@@ -6,10 +6,12 @@ test_that("get_carrier", {
   
   expect_equal(get_carrier(phone(ph_raw, "AU")),
                c(NA, "", "Optus", "Optus", "", ""))
-  # TODO - find a better example for `strict`
-  expect_equal(get_carrier(phone(ph_raw, "AU"), strict = TRUE),
-               c(NA, NA, "Optus", "Optus", "", ""))
   # TODO - find a better example for `safe`
   expect_equal(get_carrier(phone(ph_raw, "AU"), safe = TRUE),
                c(NA, NA, "", "", "", ""))
+  
+  expect_true(is.na(get_carrier(phone("0", "AU"))))
+  expect_true(is.na(get_carrier(phone("0", "AU"), strict = TRUE)))
+  expect_false(is.na(get_carrier(phone("1234", "AU"))))
+  expect_true(is.na(get_carrier(phone("1234", "AU"), strict = TRUE)))
 })